+++ dyn/core.py
@@ -78,6 +78,22 @@ class SessionEngine(Singleton):
         self._permissions = None
 
     @classmethod
+    def new_session(cls, *args, **kwargs):
+        """Return a new session instance, regardless of whether or not there is
+        already an existing session.
+
+        :param *args: Arguments to be passed to the Singleton __call__ method
+        :param *kwargs: keyword arguments to be passed to the Singleton __call__
+            method
+        """
+        cur_thread = threading.current_thread()
+        key = getattr(cls, '__metakey__')
+        instance = cls._instances.get(key, {}).get(cur_thread, None)
+        if instance:
+            instance.close_session()
+        return cls.__call__(*args, **kwargs)
+
+    @classmethod
     def get_session(cls):
         """Return the current session for this Session type or None if there is
         not an active session